Rename integration testing to e2e
Less letters to type and describes the function better. Also, we already
have ITs in some of the modules.
