+++ e2e/src/test/java/org/jdbi/v3/e2e/TestIssue2497.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.integrationtest;
+package org.jdbi.v3.e2e;
 
 import java.lang.reflect.Type;
 import java.sql.Types;
@@ -28,7 +28,7 @@ import org.jdbi.v3.core.argument.Argument;
 import org.jdbi.v3.core.argument.ArgumentFactory;
 import org.jdbi.v3.core.config.ConfigRegistry;
 import org.jdbi.v3.core.generic.GenericType;
-import org.jdbi.v3.integrationtest.TestIssue2497.Parameters.Thing;
+import org.jdbi.v3.e2e.TestIssue2497.Parameters.Thing;
 import org.jdbi.v3.sqlobject.SqlObject;
 import org.jdbi.v3.sqlobject.SqlObjectPlugin;
 import org.jdbi.v3.sqlobject.customizer.Bind;
