+++ e2e/src/test/java/org/jdbi/v3/e2e/TestIssue2508.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.integrationtest;
+package org.jdbi.v3.e2e;
 
 import java.util.ArrayList;
 import java.util.List;
